[Author: dmaclach]
Change GTMGoogleSearch to not do escaping of URL arguments, because as it turns
out the client often wants to handle that themselves. Also, made it so that
you didn't have to have a query argument, and that you could override
other arguments by setting them to [NSNull null]. These changes make the API
far more flexible.

DELTA=48 (12 added, 23 deleted, 13 changed)
R=thomasvl