[Author: thomasvl]
Fix memory model by making all the outlets assign not retain.

The problem here is UIViewControllers will dump their view in low memory
and then reload next time it is needed.  So someone hangs on to the
localizer for something else, it would keep the views also loaded, so this
way the localizer won't keep any of the objects it points to alive via a
retain.  Note: the AppKit version sorta matches this since outlets aren't
retained by default.

Clear them in awakeFromNib since that is all they are needed for.

R=dmaclach
DELTA=19  (9 added, 0 deleted, 10 changed)